+++ mail_deduplicate/deduplicate.py
@@ -42,7 +42,7 @@ from .mail_box import open_box
 if sys.version_info >= (3, 11):
     from enum import StrEnum
 else:
-    from backports.strenum import StrEnum  # type: ignore[import-not-found]
+    from backports.strenum import StrEnum
 
 TYPE_CHECKING = False
 if TYPE_CHECKING: